% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getFeaturesAllSpots}
\alias{getFeaturesAllSpots}
\title{Function used to segment all spots for a well}
\usage{
getFeaturesAllSpots(cal, seg, p)
}
\arguments{
\item{cal}{Calibrated RGB image matrix.}

\item{seg}{List of the nuclear and cell segmentation masks.}

\item{p}{List of parameters which are read from the file specified in
\code{featurePar} of the \code{extractFeatures} function.}
}
\value{
Returns a \code{data.frame} with the extracted features.
}
\description{
\code{getFeaturesAllSpots} is called by the function \code{extractFeatures}
to segment the images of the screen
}
\examples{
## see section 2.1 Image processing on cluster for a working example
localPath <- tempdir()
serverURL <- system.file("extdata", package = "PGPC")

imageConfFile = file.path("conf", "imageconf.txt")

## circumvent memory problem on 32bit windows by segementing only spot 1.
if(.Platform$OS.type == "windows" & R.Version()$arch == "i386")
    imageConfFile = file.path("conf", "imageconf_windows32.txt")

x = parseImageConf(imageConfFile,
                   localPath=localPath,
                   serverURL=serverURL)

well <- "045-01-C23"

## get segmentation parameter
p <- readHTS(x, type = "file",
             filename = file.path("conf", "segmentationpar.txt"),
             format = "dcf")

segmentation <- segmentAllSpots(x, well, p, access="cache")

## get feature parameter
pf <- readHTS(x, type = "file",
              filename = file.path("conf", "featurepar.txt"),
              format = "dcf")

ftrs <- getFeaturesAllSpots(cal=segmentation$cal,
                            seg=list(nseg=segmentation$nseg,
                                     cseg=segmentation$cseg),
                            pf)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{imageHTS}}, \code{\link{extractFeatures}}
}

