% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getInteractions}
\alias{getInteractions}
\title{Function to calculate chemical genetic interactions}
\usage{
getInteractions(d, ftrs = NULL, samplesOnly = FALSE, scaleByLine = FALSE,
  ...)
}
\arguments{
\item{d}{\code{list} containing a \code{array} with four dimensions of
features \code{D}, a \code{list} with annotation. The annotation
\code{list} needs to contain a character \code{vector} \code{ftr} with the
feature names and a character \code{vector} \code{drug$Content} defining
whether the data comes from "sample" or "other" wells.}

\item{ftrs}{Parameter to select certain features for the caluclation of
chemical genetic interactions.}

\item{samplesOnly}{If set to \code{TRUE} only values of "sample" wells will
be used for the calculation of chemical genetic interactions.}

\item{scaleByLine}{If set to \code{TRUE} the interaction terms for each
cell line will be scaled by the median absolute deviation of interaction
terms for the individual cell line and replicate.}

\item{...}{Additional parameters passed to \code{medpolish}}
}
\value{
A list with the annotation \code{anno}, raw data of the selected
features \code{D}, the chemical genetic interaction results \code{res},
the estimated drug and cell line effect \code{effect} and
calculated p-values and multiple testing adjusted p-values.
}
\description{
To detect chemcial genetic interactions, the data of each feature is
modeled using a multiplicative model and robust L1 regression to estimate
the effects of the cell line and drug treatment using the \code{medpolish}
function. In this iterative approach row and column median values are
subtracted alternately until the proportional change of the absolute
residuals falls below a defined threshold. The final row and column values
describe the drug and cell line effect respectively. The residuals
represent the interaction terms. This process is done for each replicate
and each feature individually.
To detect significant interactions the values of replicates are used to
perform a moderated t-test against the null hypothesis t = 0, using the
implementation in the Bioconductor package \code{limma}. p-values are
adjusted for multiple testing by controlling for the false discovery rate
using the method of Benjamini & Hochberg.
}
\examples{
data(interactions)
x <- getInteractions(interactions)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{medpolish}}, \code{\link{interactions}},
\code{\link{lmFit}}, \code{\link{eBayes}}
}

