% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{mergeProfiles}
\alias{mergeProfiles}
\title{Function to merge profiles of extracted features from parallel
processing}
\usage{
mergeProfiles(x, profilename = "profiles", output = "profiles.tab",
  folder = "data", access = "cache")
}
\arguments{
\item{x}{A \code{imageHTS} object.}

\item{profilename}{pattern of profile file names.}

\item{output}{File name to save the merged profiles.}

\item{folder}{folder name in which the profile files are stored and in
which the the result is saved.}

\item{access}{Access parameter passed to \code{fileHTS}}
}
\value{
None, writes the merged profiles into the specified file on disk.
}
\description{
Merges all feature profiles of each well and saves the result in the
specified file.
}
\examples{
## see section 2.1 Image processing on cluster for usage
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{imageHTS}}, \code{\link{extractFeatures}},
\code{\link{fileHTS}}
}

