% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{segmentAllSpots}
\alias{segmentAllSpots}
\title{Function used to segment all spots for a well}
\usage{
segmentAllSpots(x, uname, p, access)
}
\arguments{
\item{x}{A \code{imageHTS} object.}

\item{uname}{A character vector, containing the well names to segment. See
getUnames for details.}

\item{p}{List of parameters which are read from the file specified in
\code{segmentationPar} of the \code{segmentWells} function.}

\item{access}{A character string indicating how to access the data. Valid
values are \code{'local'}, \code{'server'} and \code{'cache'}, the default.
See \code{fileHTS} for details.}
}
\value{
Returns a \code{list} with the following items
cal: calibrated RGB image of the different channels
nseg: nuclear segmentation mask
cseg: cell segmentation mask
}
\description{
\code{segmentAllSpots} is called by the \code{imageHTS} function
\code{segmentWells} to segment the images of the screen
}
\examples{
## see section 2.1 Image processing on cluster for a working example
localPath = tempdir()
serverURL = system.file("extdata", package = "PGPC")

imageConfFile = file.path("conf", "imageconf.txt")

## circumvent memory problem on 32bit windows by segementing only spot 1.
if(.Platform$OS.type == "windows" & R.Version()$arch == "i386")
    imageConfFile = file.path("conf", "imageconf_windows32.txt")

x = parseImageConf(imageConfFile,
                   localPath=localPath,
                   serverURL=serverURL)


well = "045-01-C23"

## get segmentation parameter
p = readHTS(x, type = "file",
            filename = file.path("conf", "segmentationpar.txt"),
            format = "dcf")

segmentation = segmentAllSpots(x, well, p, access="cache")
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{imageHTS}}, \code{\link{segmentWells}},
\code{\link{fileHTS}}
}

