% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{segmentXman}
\alias{segmentXman}
\title{Function used to segment a single image}
\usage{
segmentXman(x, uname, p, access, spot = NULL)
}
\arguments{
\item{x}{A \code{imageHTS} object.}

\item{uname}{A character vector, containing the well names to segment. See
getUnames for details.}

\item{p}{List of parameters which are read from the file specified in
\code{segmentationPar} of the \code{segmentWells} function.}

\item{access}{A character string indicating how to access the data. Valid
values are \code{'local'}, \code{'server'} and \code{'cache'}, the default.
See \code{fileHTS} for details.}

\item{spot}{An single integer, indicating the spot number to segment. If it
is specified, the default \code{1} will automatically be used.}
}
\value{
Returns a \code{list} with the following items
cal: calibrated RGB image of the different channels
nseg: nuclear segmentation mask
cseg: cell segmentation mask
}
\description{
\code{segmentXman} is called by the function \code{segmentAllSpots}
to segment the single image of one spot.
}
\examples{
## see also section 2.1 "Image processing on cluster"
localPath <- tempdir()
serverURL <- system.file("extdata", package = "PGPC")

x <- parseImageConf(file.path("conf", "imageconf.txt"),
                    localPath=localPath,
                    serverURL=serverURL)

well <- "045-01-C23"

## get segmentation parameter
p <- readHTS(x, type = "file",
             filename = file.path("conf", "segmentationpar.txt"),
             format = "dcf")

segmentation <- segmentXman(x, well, p, access="cache", spot=1)
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{segmentAllSpots},\link{getUnames}}
}

