% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{summarizeWellsExtended}
\alias{summarizeWellsExtended}
\title{Function to summarize the extracted features per cell for each well}
\usage{
summarizeWellsExtended(x, uname, featurePar,
  profileFilename = file.path("data", "profiles.tab"), access = "cache")
}
\arguments{
\item{x}{A \code{imageHTS} object.}

\item{uname}{A character vector, containing the well names that will be
summarized.}

\item{featurePar}{File containing the feature parameters used for
summarizing the wells.}

\item{profileFilename}{File name to save the summarized features.}

\item{access}{Access parameter passed to fileHTS}
}
\value{
None, writes the summarized well profiles into the specified files
on disk.
}
\description{
The function extends the \code{imageHTS} function \code{summarizeWells}. It
calculates summary statistics over all cells for each well.
In particular the trimmed mean (trim = 0.1) and sd is calculated for each
extracted feature. Additionally the 1%, 5%, 95% and 99% quantiles are
calculated for features that are not the standard deviation, median absolut
deviation or Halralick statistics calculated over each cell.
}
\examples{
## see section 2.1 Image processing on cluster for a working example
localPath = tempdir()
serverURL = system.file("extdata", package = "PGPC")

imageConfFile = file.path("conf", "imageconf.txt")

## circumvent memory problem on 32bit windows by segementing only spot 1.
if(.Platform$OS.type == "windows" & R.Version()$arch == "i386")
    imageConfFile = file.path("conf", "imageconf_windows32.txt")

x = parseImageConf(imageConfFile,
                   localPath=localPath,
                   serverURL=serverURL)

well = "045-01-C23"

segmentWells(x, well, file.path("conf", "segmentationpar.txt"))
PGPC:::extractFeaturesWithParameter(x,
                                    well,
                                    file.path("conf", "featurepar.txt"))

summary <- summarizeWellsExtended(x,
                                  well,
                                  file.path("conf", "featurepar.txt"))
}
\author{
Felix A. Klein, \email{felix.klein@embl.de}
}
\seealso{
\code{\link{imageHTS}}, \code{\link{summarizeWells}}
}

